fix(callbox): loop ringtone
